Move expect_ready into libsignal-bridge
