+++ rust/bridge/shared/Cargo.toml
@@ -13,6 +13,7 @@ license = "AGPL-3.0-only"
 [dependencies]
 libsignal-protocol-rust = { path = "../../protocol" }
 aes-gcm-siv = { path = "../../aes-gcm-siv" }
+futures = "0.3.7"
 log = "0.4.11"
 paste = "1.0"
 thiserror = "1.0"
