+++ rust/bridge/shared/src/support.rs
@@ -3,7 +3,21 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 //
 
-pub use paste::paste;
+use futures::pin_mut;
+use futures::task::noop_waker_ref;
+use std::future::Future;
+use std::task::{self, Poll};
+
+pub(crate) use paste::paste;
+
+#[track_caller]
+pub fn expect_ready<F: Future>(future: F) -> F::Output {
+    pin_mut!(future);
+    match future.poll(&mut task::Context::from_waker(noop_waker_ref())) {
+        Poll::Ready(result) => result,
+        Poll::Pending => panic!("future was not ready"),
+    }
+}
 
 /// Wraps an expression in a function with a given name and type...
 /// except that if the expression is a closure with a single typeless argument,
